% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anglecheck.R
\name{anglecheck}
\alias{anglecheck}
\title{Angle swept by a ray}
\usage{
anglecheck(x, radians = TRUE)
}
\arguments{
\item{x}{a vector of length 2 giving the xy-coordinates of the point defined by the user}

\item{radians}{if FALSE, returns angle in degrees instead of radians}
}
\description{
This function computes angle swept by a ray in anti-clockwise direction.
The ray starts at the half-line pointing towards the positive x direction,
and passes through the origin and a point defined by the user.
}
\examples{
anglecheck(c(0,1),radians=FALSE)
}
\author{
Tsung Fei Khang \email{tfkhang@um.edu.my}
}
\references{
Khang TF, Soo OYM, Tan WB, Lim LHS. (2016). Monogenean anchor morphometry: systematic value, phylogenetic signal, and evolution. PeerJ 4:e1668.
}

