% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorBar.R
\name{colorBar}
\alias{colorBar}
\title{Add a color bar}
\usage{
colorBar(colpalette = NULL, min, max = -min, nticks = 11,
  ticks = seq(min, max, len = nticks), tit = "")
}
\arguments{
\item{colpalette}{a color palette supplied by the user; defaults to the red-white-blue palette if left undefined}

\item{min}{the smallest numeric value corresponding to the color on the extreme left of the color palette}

\item{max}{the largest numeric value corresponding to the color on the extreme right of the color palette}

\item{nticks}{number of tick lines on the color bar}

\item{ticks}{spacing for the tick marks on the color bar}

\item{tit}{title for the color bar}
}
\description{
This function produces a color bar.
}
\details{
This R code is a based on John Colby's (2011) \code{color.bar} code. A user-defined color palette can be generated using the
\code{colorRampPalette} function. See www.colorbrewer2.org for interesting color palette options.
}
\examples{
#define a color scale
tones <- c("#99000D", "#FB6A4A", "white", "#6BAED6", "#084594")
multihue <- colorRampPalette(tones, space = "Lab")(101)

#create a two-panel figure with width ratio of 4:1
nf <- layout(matrix(c(1,1,1,1,2),2,5,byrow=TRUE))
layout.show(nf)

#mapping a matrix of randomly chosen numbers between -1 and 1 to
#colors in the color scale
h <- matrix(runif(100, -1, 1),10,10)
image(h,col=multihue,xaxt="n", yaxt="n")

#add color bar
par(mar=c(5,5,5,2))
colorBar(colpalette=multihue, min=-1,max=1)
}
\author{
Tsung Fei Khang \email{tfkhang@um.edu.my}
}
\references{
Colby J. (2011). Color bar legends for neuroimaging in R. Available at http://www.colbyimaging.com/wiki/statistics/color-bars.

Khang TF, Soo OYM, Tan WB, Lim LHS. (2015). Monogenean anchor morphometry: systematic value, phylogenetic signal and evolution.
}
\seealso{
\code{\link{pcloadhm}}, \code{\link{colorRampPalette}}
}

