% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spcolmap.R
\docType{data}
\name{spcolmap}
\alias{spcolmap}
\title{Species-color map}
\format{a matrix - the first column gives the species names; the second their corresponding color code; the third their host species}
\source{
Khang TF, Soo OYM, Tan WB, Lim LHS. (2015). Monogenean anchor morphometry: systematic value, phylogenetic signal and evolution.
}
\usage{
data(spcolmap)
}
\description{
Guide for species coloring in plots.
}
\author{
Tsung Fei Khang \email{tfkhang@um.edu.my}
}
\keyword{datasets}

