% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tpsRead.R
\name{tpsRead}
\alias{tpsRead}
\title{Reading tps files into R}
\usage{
tpsRead(x)
}
\arguments{
\item{x}{a constant that indexes the folder of interest}
}
\value{
a list of matrices containing the xy-coordinates of landmarks
}
\description{
This function converts all tps files in a folder to a list of matrices. The tps files are created
using the TPSDIG2 software (Rohlf, 2013, 2015).
}
\examples{
##not run
##download the tps data folders from Khang et al. (2015) and unzip them into a folder called ligotps

#setwd("ligotps/")
#dir()
##read tps files from the /bantingensis folder and turn them into objects of a list
#xydata <- tpsRead(1)
}
\author{
Tsung Fei Khang \email{tfkhang@um.edu.my}
}
\references{
Khang TF, Soo OYM, Tan WB, Lim LHS. (2015). Monogenean anchor morphometry: systematic value, phylogenetic signal and evolution.

Rohlf FJ. (2013). Morphometrics at SUNY Stony Brook. Available at http://life.bio.sunysb.edu/morph/soft-dataacq.html.

Rohlf FJ. (2015). The tps series of software. Hystrix. 26, doi: 10.4404/hystrix-26.1-11264.
}

