% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwed_pd.R
\docType{data}
\name{pwed_pd}
\alias{pwed_pd}
\title{Physical distance between landmarks in ventral and dorsal anchors of 13 \emph{Ligophorus} species}
\format{a matrix containing 437 rows (samples) and 110 columns (physical distance)}
\source{
Khang TF, Soo OYM, Tan WB, Lim LHS. (2016). Monogenean anchor morphometry: systematic value, phylogenetic signal, and evolution. PeerJ 4:e1668.
}
\usage{
data(pwed_pd)
}
\description{
This data set contains all pairwise physical distance between landmarks in
ventral and dorsal anchors of samples from 13 \emph{Ligophorus species}.
}
\details{
The pairwise physical distances are estimated from a subset (n=96) of the 437 samples based on a regression equation,
using physical distances between landmarks 1 to 3 and landmarks 1 to 5 and corresponding distances computed from \code{pwdist},
which have arbitrary units.
}
\author{
Tsung Fei Khang \email{tfkhang@um.edu.my}
}
\seealso{
\code{\link{pwdist}}
}
\keyword{datasets}

